+++ CHANGELOG.md
@@ -4,12 +4,22 @@ We are currently working on porting this changelog to the specifications in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Version 0.2.0] - 
+## [Version 0.2.1] - 
+
+### Added
+* autoconf-gpg for auto-configuring which gpg key to sign commits with
+
+### Changed
+* git track-upstream will now ask which remote to use when it is ambiguous.
+* Improved `git-remote-protocol`
+
+## [Version 0.2.0] - Released 2023-08-09
 
 ### Added
 * Add `git_rebase_add_continue`
+* Add `git_remote_protocol` (i.e. git permit)
 
-## [Version 0.1.1] - Released 2022-06-22
+## [Version 0.1.1] - Released 2023-06-22
 
 ### Added
 * Initial version with `branch_upgrade`,`squash_streaks`,`sync`,`branch_cleanup`,`track_upstream`